split tests to diff files
