+++ lightgbm/src/test/scala/com/microsoft/azure/synapse/ml/lightgbm/split1/VerifyLightGBMClassifier.scala
@@ -8,7 +8,6 @@ import com.microsoft.azure.synapse.ml.core.test.fuzzing.{EstimatorFuzzing, TestO
 import com.microsoft.azure.synapse.ml.lightgbm._
 import com.microsoft.azure.synapse.ml.lightgbm.dataset.{ChunkedArrayUtils, LightGBMDataset}
 import com.microsoft.azure.synapse.ml.lightgbm.params.FObjTrait
-import com.microsoft.azure.synapse.ml.lightgbm.swig.{DoubleChunkedArray, DoubleSwigArray, SwigUtils}
 import com.microsoft.azure.synapse.ml.stages.MultiColumnAdapter
 import org.apache.commons.io.FileUtils
 import org.apache.spark.TaskContext
@@ -184,69 +183,6 @@ class VerifyLightGBMClassifier extends Benchmarks with EstimatorFuzzing[LightGBM
     val scoredDF2 = baseModel.setInitScoreCol(initScoreCol).fit(df2).transform(df2)
 
     assertBinaryImprovement(scoredDF1, scoredDF2)
-  }
-
-  test("Verify chunked array transpose simple") {
-    Array(10, 100).foreach(chunkSize => {
-      LightGBMUtils.initializeNativeLibrary()
-      val rows = 10
-      val cols = 2
-      val chunkedArray = new DoubleChunkedArray(chunkSize) // either whole chunks or 1 incomplete chunk
-      val transposedArray = new DoubleSwigArray(rows * cols)
-
-      // Create transposed array (for easier validation since transpose will convert to sequential)
-      for (row <- 0L until rows) {
-        for (col <- 0L until cols) {
-          chunkedArray.add(row + rows * col)
-        }
-      }
-
-      try {
-        ChunkedArrayUtils.insertTransposedChunkedArray(chunkedArray, cols, transposedArray, rows, 0)
-
-        // Assert row order in source (at least for first row)
-        assert(chunkedArray.getItem(0, 0, 0) == 0)
-        assert(chunkedArray.getItem(0, 1, 0) == rows)
-
-        // Assert column order in source (should be sequential numbers)
-        val array = SwigUtils.nativeDoubleArrayToArray(transposedArray.array, rows * cols)
-        assert(array.zipWithIndex.forall(pair => pair._1 == pair._2))
-      } finally {
-        transposedArray.delete()
-      }
-    })
-  }
-
-  test("Verify chunked array transpose complex") {
-    LightGBMUtils.initializeNativeLibrary()
-    val rows = 10
-    val cols = 2
-    val chunkedArray = new DoubleChunkedArray(7) // ensure partial chunks
-    val transposedArray = new DoubleSwigArray(rows * cols * 2)
-    for (row <- 0L until rows) {
-      for (col <- 0L until cols) {
-        chunkedArray.add(row + rows * col)
-      }
-    }
-
-    try {
-      // copy into start and middle
-      ChunkedArrayUtils.insertTransposedChunkedArray(chunkedArray, cols, transposedArray, rows * 2, 0)
-      ChunkedArrayUtils.insertTransposedChunkedArray(chunkedArray, cols, transposedArray, rows * 2, rows)
-
-      // Assert row order in source (at least for first row)
-      assert(chunkedArray.getItem(0, 0, 0) == 0)
-      assert(chunkedArray.getItem(0, 1, 0) == rows)
-
-      val array = SwigUtils.nativeDoubleArrayToArray(transposedArray.array, rows * cols * 2)
-      val expectedArray = ((0 until rows)
-                          ++ (0 until rows)
-                          ++ (rows until 2*rows)
-                          ++ (rows until 2*rows))
-      assert(array.zipWithIndex.forall(pair => pair._1 == expectedArray(pair._2)))
-    } finally {
-      transposedArray.delete()
-    }
   }
 
   test("Verify LightGBM Multiclass Classifier with vector initial score") {
